test_sdcard.sh: modify script to add continuous mode support

+++ src/test_sdcard.sh
@@ -15,20 +15,56 @@
 # And it will work only if sdcard has fat filesystem
 
 LOG_LEVEL=1
+TRIALS=1
+CONTINUOUS=false
 
 source common.sh
-parse_command_line $@
+
+usage()
+{
+cat << EOF
+
+usage: $0 options
+
+OPTIONS:
+-h	Show this message
+-c	Number of trials, default 1, and pass -c 0 for continuous mode
+-v	Verbose
+-V	Show package version
+
+EOF
+}
+
+while getopts "c:vVh" opt; do
+	case $opt in
+		c)
+			TRIALS=$OPTARG;;
+		v)
+			LOG_LEVEL=2;;
+		V)
+			echo -n "version = "
+			cat version
+			exit 0;;
+		h)
+			usage
+			exit 0;;
+		\?)
+			usage
+			exit 1;;
+	esac
+done
+
 redirect_output_and_error $LOG_LEVEL
 
 echo -e "\n******************************* SD card test **********************************\n" >&3
 
-COUNT=`ls /dev | grep mmcblk0p -c`
-TOTAL_MMC_DEVS=$COUNT
+TOTAL_MMC_DEVS=`ls /dev | grep mmcblk0p -c`
 
-if [ $COUNT -ne 0 ];then
+echo "$COUNT mmc partitions found"
 
-	echo "$COUNT mmc partitions found"
-
+create_mountpoints()
+{
+	COUNT=$1
 	while [ $COUNT -gt 0 ]
 	do
 		COUNT=`expr $COUNT - 1`
@@ -41,16 +77,44 @@ if [ $COUNT -ne 0 ];then
 			rm -rf $MOUNT_DIR
 		fi
 
-		mkdir -p $MOUNT_DIR		&&\
-		mount -t vfat /dev/$MMC_DEVICE $MOUNT_DIR	&&\
-		dd if=/dev/zero of=/tmp/temp0.img bs=1k count=1024		&&\
-		dd if=/tmp/temp0.img of=$MOUNT_DIR/temp1.img	&&\
-		cmp /tmp/temp0.img $MOUNT_DIR/temp1.img && COUNT=0
-		{
-			[ $? == 0 ] && echo "PASS" || (echo "FAIL - partition $MMC_DEVICE"; if [ $COUNT = 1 ]; then exit 1; fi)
-		} >&3
+		mkdir -p $MOUNT_DIR
+		mount -t vfat /dev/$MMC_DEVICE $MOUNT_DIR
+	done
+}
 
-		rm /tmp/temp0.img $MOUNT_DIR/temp1.img >&4
+if [ $TOTAL_MMC_DEVS -ne 0 ];then
+
+	create_mountpoints $TOTAL_MMC_DEVS
+
+	#check if we are asked to run continuously
+	if [ $TRIALS -eq 0 ];then
+		CONTINUOUS=true
+	fi
+
+	while [ $CONTINUOUS == true -o $TRIALS -gt 0 ]
+	do
+		COUNT=$TOTAL_MMC_DEVS
+		while [ $COUNT -gt 0 ]
+		do
+			COUNT=$((COUNT-1))
+			NUM=$(($TOTAL_MMC_DEVS-$COUNT))
+			MMC_DEVICE=`ls /dev/ | grep mmcblk0p | head -n $NUM | tail -n 1`
+			MOUNT_DIR=/mnt/$MMC_DEVICE
+
+			if mountpoint -q -- $MOUNT_DIR;then
+				dd if=/dev/urandom of=/tmp/temp0.img bs=1k count=1024	&&\
+				cp /tmp/temp0.img $MOUNT_DIR/temp1.img	&&\
+				sync	&&\
+				cmp /tmp/temp0.img $MOUNT_DIR/temp1.img
+				{
+					[ $? == 0 ] && echo "PASS" || (echo "FAIL - partition $MMC_DEVICE"; if [ $COUNT == 0 ]; then exit 1; fi)
+				}
+
+				rm /tmp/temp0.img $MOUNT_DIR/temp1.img >&4
+			fi
+		done
+		TRIALS=$(($TRIALS-1))
+		sleep 1
 	done
 
 else
